requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
biip==4.0.0
certifi==2025.6.15
chardet==5.2.0
coverage==7.8.0
crispy-bootstrap5==2025.4
distro==1.9.0
Django==5.2.1
django-appconf==1.1.0
django-bootstrap5==25.1
django-crispy-forms==2.4
django-filter==25.1
django-htmx==1.23.0
django-redis==5.4.0
django-select2==8.4.0
django-tables2==2.7.5
django-template-partials==24.4
factory_boy==3.3.3
Faker==37.3.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
narwhals==1.39.1
numpy
openai==1.93.0
packaging==25.0
pandas
pdf2image==1.17.0
pillow
plotly==6.1.0
pluggy==1.6.0
psycopg==3.2.9
pydantic
pydantic_core
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==6.2.0
reportlab==4.4.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tablib==3.8.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
whitenoise==6.9.0
zxing-cpp
